% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdjhR.R
\name{pyrdjhmain}
\alias{pyrdjhmain}
\title{Title 历史应收票据汇票等级更新}
\usage{
pyrdjhmain(input_excel, token = "F91CF3E3-8962-47F2-823F-C5CCAAFC66CA")
}
\arguments{
\item{input_excel}{文件路径}

\item{token}{口令}
}
\value{
输出结果:"更新完成"
}
\description{
Title 历史应收票据汇票等级更新
}
\examples{
pyrdjhmain(input_excel='需处理历史单据清单.xlsx', token='F91CF3E3-8962-47F2-823F-C5CCAAFC66CA')

}
