% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdjhR.R
\name{pyrdjhReadData}
\alias{pyrdjhReadData}
\title{Title 读取文件}
\usage{
pyrdjhReadData(input_excel)
}
\arguments{
\item{input_excel}{输入文件路径}
}
\value{
输出datafram
}
\description{
Title 读取文件
}
\examples{
pyrdjhReadData('需处理历史单据清单.xlsx')

}
