% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdjhR.R
\name{pyrdjhres}
\alias{pyrdjhres}
\title{Title 查阅更新结果}
\usage{
pyrdjhres(token = "F91CF3E3-8962-47F2-823F-C5CCAAFC66CA")
}
\arguments{
\item{token}{口令}
}
\value{
返回更新结果
}
\description{
Title 查阅更新结果
}
\examples{
pyrdjhres()

}
