% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdjhR.R
\name{pyrdjhUpdateDraftGrad}
\alias{pyrdjhUpdateDraftGrad}
\title{Title 更新汇票等级}
\usage{
pyrdjhUpdateDraftGrad(token, df)
}
\arguments{
\item{token}{口令}

\item{df}{datafram数据}
}
\value{
返回更新结果:更新成功
}
\description{
Title 更新汇票等级
}
\examples{
pyrdjhUpdateDraftGrad(token=F91CF3E3-8962-47F2-823F-C5CCAAFC66CA, df=pyrdjhReadData('需处理历史单据清单.xlsx'))
}
